Split ragfair and settings code from server
