+++ src/settings.js
@@ -1,0 +1,5 @@
+var settings = JSON.parse(utility.readJson("settings.json"));
+
+function getPort() {
+	return settings.server.port;
+}
